armor-bench-starter/requirements.txt
transformers>=4.43
datasets>=2.19
sentence-transformers>=2.6
scikit-learn>=1.4
pandas>=2.2
numpy>=1.26
PyYAML>=6.0
tqdm>=4.66
matplotlib>=3.8
python-docx>=1.1
pdfminer.six>=20231228
